cloud_journey/src/requirements.txt
#cloud_journey/src/requirements.txt
djangodocker compose run app django-admin startproject sneafur .